+++ querki/scalajs/src/main/scala/querki/time/package.scala
@@ -2,7 +2,7 @@ package querki
 
 import scala.scalajs.js
 
-import org.widok.moment._
+import moment._
 
 import querki.globals._
 
@@ -11,7 +11,7 @@ package object time {
   // HACK to get common time formatting. In the medium term this needs to become much smarter,
   // to respond correctly to localization.
   def displayTime(time: Common.Timestamp): String = {
-    Moment(time).calendar(
+    Moment(time.toDouble).calendar(
       js.undefined,
       lit(
         sameElse = "MM/DD/YYYY"
